requirements.txt
aiohttp==3.5.4
alembic==1.0.10
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.0.1
idna==2.8
isort==4.3.18
jellyfish==0.7.1
lazy-object-proxy==1.3.1
Mako==1.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
psycopg2-binary==2.8.2
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
websockets==6.0
wrapt==1.11.1
yarl==1.3.0
